[Feat] 1. 유저뱃지 목록 조회 2. 찜 취소/재찜 구현 3. 팔로우/언팔로우 구현 4. 팔로우 창 팝업으로 변경

+++ back-end/foodiefair/src/main/java/com/multicampus/foodiefair/controller/VisitedController.java
@@ -18,7 +18,7 @@ public class VisitedController {
 
     @GetMapping("/visited")
     public ResponseEntity<VisitedPageResponseDTO> getVisitedList(
-            @PathVariable Long userId,
+            @PathVariable int userId,
             @RequestParam(required = false, defaultValue = "1") int page,
             @RequestParam(required = false, defaultValue = "5") int size) {
