+++ back-end/foodiefair/src/main/java/com/multicampus/foodiefair/dao/FollowDAO.java
@@ -7,15 +7,15 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface FollowDAO {
     // 다른 유저의 자신 팔로우 취소
-    public int deleteFollowed(Long followingId, Long followedId);
+    public int deleteFollowed(int followingId, int followedId);
     // 유저 팔로우
     public int insertFollowing(FollowDTO followDTO);
     // 유저 언팔로우
-    public int deleteFollowing(Long followingId, Long followedId);
+    public int deleteFollowing(int followingId, int followedId);
     // 나를 Follow하는 유저 수
-    public int selectFollowerCount(Long followedId);
+    public int selectFollowerCount(int followedId);
     // 내가 Follow하는 유저 수
-    public int selectFollowingCount(Long followingId);
+    public int selectFollowingCount(int followingId);
     // 팔로우 상태 확인 (내가 팔로우 중인지)
-    public Boolean check(Long followingId, Long followedId);
+    public Boolean check(int followingId, int followedId);
 }
