+++ back-end/foodiefair/src/main/java/com/multicampus/foodiefair/service/UserTagService.java
@@ -4,6 +4,6 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 
 public interface UserTagService {
-    String getUserTags(Long userId);
-    String getUserBadges(Long userId);
+    String getUserTags(int userId);
+    String getUserBadges(int userId);
 }
